Conflicts:
	api/tasks.py
	api/test/test_tasks.py
	api/views.py
	ui/app/assets/partials/testresults/test_details.html
	ui/app/scripts/directives.coffee
	ui/app/scripts/testresults/controllers.coffee
	ui/app/scripts/testresults/model.coffee

+++ api/models.py
@@ -17,6 +17,13 @@ from api.amazon_utils import AmazonS3Helper
 
 @app.conn.register
 class LogMessage(Document):
+    TRAIN_MODEL = 'trainmodel_log'
+    IMPORT_DATA = 'importdata_log'
+    RUN_TEST = 'runtest_log'
+    CONFUSION_MATRIX_LOG = 'confusion_matrix_log'
+    TYPE_CHOICES = (TRAIN_MODEL, IMPORT_DATA, RUN_TEST,
+                    CONFUSION_MATRIX_LOG)
+
     __collection__ = 'logs'
     structure = {
         # error, warning
@@ -28,6 +35,7 @@ class LogMessage(Document):
         'created_on': datetime,
     }
     default_values = {'created_on': datetime.utcnow}
+    use_dot_notation = True
 
     @classmethod
     def delete_related_logs(cls, obj, level=None):
@@ -151,7 +159,7 @@ class DataSet(Document):
     __collection__ = 'dataset'
     LOG_TYPE = 'importdata_log'
 
-    STATUS_IMPORTINING = 'Importing'
+    STATUS_IMPORTING = 'Importing'
     STATUS_IMPORTED = 'Imported'
     STATUS_ERROR = 'Error'
 
@@ -167,6 +175,10 @@ class DataSet(Document):
         'on_s3': bool,
         'compress': bool,
         'filename': basestring,
+        'filesize': long,
+        'records_count': int,
+        'time': int,
+        'data_fields': list,
     }
     required_fields = ['name', 'created_on', 'updated_on', ]
     default_values = {'created_on': datetime.utcnow,
@@ -174,7 +186,8 @@ class DataSet(Document):
                       'error': '',
                       'on_s3': False,
                       'compress': True,
-                      'status': STATUS_IMPORTINING}
+                      'status': STATUS_IMPORTING,
+                      'data_fields': []}
     use_dot_notation = True
 
     def __init__(self, *args, **kwargs):
@@ -185,20 +198,29 @@ class DataSet(Document):
         return helper.get_download_url(str(self._id), expires_in)
 
     def set_file_path(self):
-        self.data = '%s.%s' % (self._id, 'gz' if self.compress else 'json')
+        data = '%s.%s' % (self._id, 'gz' if self.compress else 'json')
+        self.data = data
         path = app.config['DATA_FOLDER']
         if not exists(path):
             makedirs(path)
-        self.filename = join(path, self.data)
+        self.filename = join(path, data)
         self.save()
+
+    @property
+    def data(self):
+        if not self.on_s3:
+            raise Exception('Invalid oper')
+
+        if not hasattr(self, '_data'):
+            self._data = self.load_from_s3()
+        return self._data
 
     def get_data_stream(self):
         import gzip
         #import zlib
         if self.on_s3:
             logging.info('Loading data from Amazon S3')
-            data = self.load_from_s3()
-            stream = StringIO.StringIO(data)
+            stream = StringIO.StringIO(self.data)
             if self.compress:
                 logging.info('Decompress data')
                 return gzip.GzipFile(fileobj=stream, mode='r')
@@ -300,11 +322,14 @@ class Model(Document):
 
     STATUS_NEW = 'New'
     STATUS_QUEUED = 'Queued'
-    STATUS_IMPORTINING = 'Importing'
+    STATUS_IMPORTING = 'Importing'
     STATUS_IMPORTED = 'Imported'
+    STATUS_REQUESTING_INSTANCE = 'Requesting Instance'
+    STATUS_INSTANCE_STARTED = 'Instance Started'
     STATUS_TRAINING = 'Training'
     STATUS_TRAINED = 'Trained'
     STATUS_ERROR = 'Error'
+    STATUS_CANCELED = 'Canceled'
 
     __collection__ = 'models'
     structure = {
@@ -338,6 +363,9 @@ class Model(Document):
         'example_label': basestring,
         'example_id': basestring,
         'tags': list,
+
+        'spot_instance_request_id': basestring,
+        'memory_usage': dict,
     }
     gridfs = {'files': ['trainer']}
     required_fields = ['name', 'created_on', ]
@@ -346,7 +374,10 @@ class Model(Document):
                       'status': STATUS_NEW,
                       'comparable': False,
                       'tags': [],
-                      'weights_synchronized': False}
+                      'weights_synchronized': False,
+                      'spot_instance_request_id': '',
+                      'memory_usage': {},
+                      }
     use_dot_notation = True
     use_autorefs = True
 
@@ -370,13 +401,15 @@ class Model(Document):
         trainer = self.get_trainer()
         fp = dataset.get_data_stream()
         try:
-            metrics = trainer.test(streamingiterload(fp), callback=callback)
+            metrics = trainer.test(
+                streamingiterload(fp),
+                callback=callback,
+                save_raw=False)
         finally:
             fp.close()
-
-        raw_data = trainer._raw_data
+        # raw_data = trainer._raw_data
         trainer.clear_temp_data()
-        return metrics, raw_data
+        return metrics
 
     def set_trainer(self, trainer):
         from core.trainer.store import TrainerStorage
@@ -398,8 +431,9 @@ class Model(Document):
         if commit:
             self.save()
 
-    def delete_metadata(self):
-        LogMessage.delete_related_logs(self)
+    def delete_metadata(self, delete_log=True):
+        if delete_log:
+            LogMessage.delete_related_logs(self)
         params = {'model_id': str(self._id)}
         app.db.Test.collection.remove(params)
         app.db.TestExample.collection.remove(params)
@@ -412,11 +446,14 @@ class Test(Document):
     LOG_TYPE = 'runtest_log'
 
     STATUS_QUEUED = 'Queued'
-    STATUS_IMPORTINING = 'Importing'
+    STATUS_IMPORTING = 'Importing'
     STATUS_IMPORTED = 'Imported'
     STATUS_IN_PROGRESS = 'In Progress'
     STATUS_COMPLETED = 'Completed'
     STATUS_ERROR = 'Error'
+
+    EXPORT_STATUS_IN_PROGRESS = 'In Progress'
+    EXPORT_STATUS_COMPLETED = 'Completed'
 
     __collection__ = 'tests'
     structure = {
@@ -438,12 +475,18 @@ class Test(Document):
         'dataset': DataSet,
         # Raw test data
         #'examples': [TestExample ],
+        'memory_usage': dict,
+        'exports': list,
     }
     required_fields = ['name', 'created_on', 'updated_on',
                        'status']
-    default_values = {'created_on': datetime.utcnow,
-                      'updated_on': datetime.utcnow,
-                      'status': STATUS_QUEUED}
+    default_values = {
+        'created_on': datetime.utcnow,
+        'updated_on': datetime.utcnow,
+        'status': STATUS_QUEUED,
+        'memory_usage': {},
+        'exports': []
+    }
     use_dot_notation = True
     use_autorefs = True
 
@@ -463,10 +506,51 @@ class Test(Document):
         self.collection.remove({'_id': self._id})
         LogMessage.delete_related_logs(self)
 
+    def get_examples_full_data(self, fields=None, data_input_params=None):
+        """
+        Get examples with data_input fields from dataset stored at s3
+        :param fields: list of needed examples fields
+        :return: iterator
+        """
+        example_id_field = self.model.example_id
+        dataset_data_stream = self.dataset.get_data_stream()
+
+        if fields:
+            for required_field in ['_id', 'id', example_id_field]:
+                if required_field not in fields:
+                    fields.append(required_field)
+
+        examples_data = app.db.TestExample.find(
+            {'model_id': self.model_id, 'test_id': str(self._id)},
+            fields
+        )
+        examples_data = dict([(epl['id'], epl)
+                              for epl in examples_data])
+
+        for row in dataset_data_stream:
+            data = json.loads(row)
+            example_id = data[example_id_field]
+            example = examples_data.get(example_id)
+            if not example:
+                continue
+            for key in data:
+                new_key = 'data_input.{0}'.format(key.replace('.', '->'))
+                example[new_key] = data[key]
+            if data_input_params:
+                skip_it = False
+                for k, v in data_input_params.items():
+                    if example[k] != v:
+                        skip_it = True
+                        break
+                if skip_it:
+                    continue
+            yield example
+
 
 @app.conn.register
 class TestExample(Document):
     __collection__ = 'example'
+    S3_KEY = 'text_example_'
 
     structure = {
         'created_on': datetime,
@@ -491,6 +575,50 @@ class TestExample(Document):
     use_autorefs = True
     default_values = {'created_on': datetime.utcnow}
     required_fields = ['created_on', ]
+    use_dot_notation = True
+
+    @property
+    def s3_key(self):
+        return '{0!s}_{1!s}'.format(self.S3_KEY, self._id)
+
+    def _save_to_s3(self, data):
+        meta = {
+            'example_id': self.id,
+            'test_id': self.test_id,
+            'model_id': self.model_id,
+        }
+        helper = AmazonS3Helper()
+        helper.save_key_string(self.s3_key, json.dumps(data), meta)
+        helper.close()
+
+    def _load_from_s3(self):
+        helper = AmazonS3Helper()
+        return helper.load_key(self.s3_key)
+
+    @property
+    def data_input(self):
+        if not self['data_input'] and getattr(self, '_id'):
+            data = self._load_from_s3()
+            self['data_input'] = json.loads(data)
+        return self['data_input']
+
+    @data_input.setter
+    def set_data_input(self, value):
+        self['data_input'] = value
+
+    def save(self, *args, **kwargs):
+        data_input = None
+        if self['data_input']:
+            data_input = self['data_input']
+            self['data_input'] = None
+        super(TestExample, self).save(*args, **kwargs)
+        if data_input:
+            self._save_to_s3(data_input)
+
+    def delete(self):
+        helper = AmazonS3Helper()
+        helper.delete_key(self.s3_key)
+        super(TestExample, self).delete()
 
 
 @app.conn.register
